+++ RangePicker/tests/RangePicker-specs.js
@@ -2,45 +2,62 @@ import {mount} from 'enzyme';
 
 import {RangePicker, RangePickerBase} from '../RangePicker';
 
-import css from '../../internal/Picker/Picker.module.less';
+import css from '../../internal/DrumPicker/DrumPicker.module.less';
 
-const decrement = (picker) => picker.find(`.${css.itemDecrement}`).first().simulate('click');
-const increment = (picker) => picker.find(`.${css.itemIncrement}`).first().simulate('click');
+const keyDown = (keyCode) => (picker) => picker.find(`.${css.root}`).first().simulate('keydown', {keyCode});
+
+const upKeyDown = keyDown(38);
+const downKeyDown = keyDown(40);
 
 describe('RangePicker Specs', () => {
+	beforeEach(() => {
+		global.Element.prototype.getBoundingClientRect = jest.fn(() => {
+			return {
+				bottom: 310,
+				height: 84,
+				left: 45,
+				right: 1348,
+				top: 226,
+				width: 1303,
+				x: 45,
+				y: 226
+			};
+		});
+	});
+
 	test('should render a single child with the current value', () => {
 		const picker = mount(
 			<RangePicker min={-10} max={20} value={10} />
 		);
 
 		const expected = '10';
-		const actual = picker.find('.active').first().text();
+		const actual = picker.find('.selectedItem').first().text();
 
 		expect(actual).toBe(expected);
 	});
 
 	test('should increase by step amount on increment press', () => {
 		const picker = mount(
-			<RangePicker defaultValue={10} min={0} max={100} noAnimation step={1} />
+			<RangePicker defaultValue={10} min={0} max={20} noAnimation step={1} />
 		);
 
-		increment(picker);
+		downKeyDown(picker);
 
 		const expected = '11';
-		const actual = picker.find('.active').first().text();
+		const actual = picker.find('.selectedItem').first().text();
 
 		expect(actual).toBe(expected);
 	});
 
 	test('should decrease by step amount on decrement press', () => {
 		const picker = mount(
-			<RangePicker defaultValue={10} min={0} max={100} noAnimation step={1} />
+			<RangePicker defaultValue={10} min={0} max={20} noAnimation step={1} />
 		);
 
-		decrement(picker);
+		upKeyDown(picker);
 
 		const expected = '9';
-		const actual = picker.find('.active').first().text();
+		const actual = picker.find('.selectedItem').first().text();
 
 		expect(actual).toBe(expected);
 	});
@@ -50,7 +67,7 @@ describe('RangePicker Specs', () => {
 			<RangePickerBase min={0} max={0} value={0} />
 		);
 
-		const actual = picker.find('Picker').last().prop('disabled');
+		const actual = picker.find('DrumPicker').last().prop('disabled');
 		expect(actual).toBe(true);
 	});
 });